+++ content/4-person-feedback-loop/description.txt
@@ -1,1 +1,3 @@
-description
+my friend <a href="https://soundcloud.com/lukasipo">lukas</a> and I were fucking around with feedback routing between computers over the internet using a software for live audio streaming called <a href="https://www.sonobus.net/">sonobus</a>
+        it was really fun and quite funny, so we got some other people to join in, manifesting in a feedback loop between 7 people, each with different processing chains on the audio.
+        In terms of routing, i'm pretty sure this could be modelled as a 7x7 feedback delay network, where each channel feeds into the six others but not directly into itself. see <a href="https://ccrma.stanford.edu/~jos/pasp/Feedback_Delay_Networks_FDN.html">this explanation of FDNs</a> for context
