+++ content/Resen/data.json
@@ -16,5 +16,9 @@
     },
     "contributors": {
         "Popbot": "primary creator"
+    },
+    "media": {
+        "resen-cover.png": "cover artwork i made for resen, made in blender"
     }
+
 }
