+++ content/Resen/description.txt
@@ -1,1 +1,2 @@
-<h1>do this</h1>
+I remember my inspiration for this track was very specific. I remember when i was walking home from school one day, a particularly plump water droplet rolled off of a leaf and fell onto my nose and splashed into my eye. That particular day at school had been a rough one, and this singular moment was able to lighten my mood because it somehow made me think in a dferent way.
+I wanted to try to make a song inspired by this moment. I was listening a lot to RAWT and Frequent's dream recall album, which both had an influence on the sound of this track. I remember expermenting a lot with granular sound design, particularly with a m4l device that at the time was new in live 11 called pitchloop89. This device is a really cool sorta granular delay, and is able to create really really neat noises. Almost all of the sound design in the track comes from this effect.
