+++ content/big-flourish/data.json
@@ -1,0 +1,22 @@
+{
+    "title": "Big Flourish",
+    "dates": {
+        "yyyy-mm-dd": "published"
+    },
+    "categories": [
+        "music"
+    ],
+    "tags": [
+        "collab",
+        "megacollab",
+        "single"
+    ],
+    "links": {
+        "soundcloud": "https://soundcloud.com/popbot1/sets/keitha-manstretta-popbot?si=0d3a210a0b924d1eaff2f32cedae93c5&utm_source=clipboard&utm_medium=text&utm_campaign=social_sharing",
+        "bandcamp": "https://artbyform.bandcamp.com/track/big-flourish",
+        "webpage": "./content/big-flourish/index.html"
+    },
+    "contributors": {
+        "Popbot": "primary creator"
+    }
+}
