+++ content/big-flourish/description.txt
@@ -1,0 +1,2 @@
+All Nighter is an event where producers are given 24 hours to write a song. Some of those songs are then chosen to be included on an official compilation. For this specific even there were ~570 submitted tracks.
+This track was made for that.
