+++ content/big-flourish/index.html
@@ -1,0 +1,36 @@
+<html>
+    <head>
+                                        <!-- Google tag (gtag.js) -->
+        <script async src="https://www.googletagmanager.com/gtag/js?id=G-ZMF5SZ58X2"></script>
+        <script>
+            window.dataLayer = window.dataLayer || [];
+            function gtag(){dataLayer.push(arguments);}
+            gtag('js', new Date());
+
+            gtag('config', 'G-ZMF5SZ58X2');
+        </script>
+    </head>
+    <body>
+        <a href="../index.html">home</a><br><br>
+        <p>a megacollab project i worked on through discord for the SVS competetion. led by <a href="https://soundcloud.com/stasysmusial" target="_blank">Stasys</a>. 
+            My contributions are during the last third of the first track. I played marimba, and arranged a pseudo "drop" bit full of more orchestral/organic sounds. 
+            I'm not entirely happy with my section because it was incredibly rushed because of procrastination.
+            Happy to have worked with a great group of people tho! It taught me a bit about collaboration styles and habits. 
+            One interesting thing was that instead of passing project files back and forth, stasys had just a single project file for everything. 
+            He wrote all the melodic and general rhythmic content, split each section up for designated peopole, gave each person a fixed amt of time to work on it, and then re-assembled the bits himself and did a bunch of cohesion and mixing work to get it sounding uniform. 
+            it's a really neat way of working, and has def inspired a current project (as of april 6th 2024). Hopefully it's a past project by the time somebody is reading this. 
+        idk why anybody would read this tbh. i'm just yapping. nothing of use is in this text. uhh. i mean i guess if you're reading this then i have you're attention. hmmm. my paypal address is legowoodtrashcan@gmail.com. give me money. </p>
+        
+        <a href="https://soundcloud.com/stasysmusial/sets/through-sharp-eyes-we-witness" target="_blank">soundcloud</a><br>
+        <a href="https://stasysmusial.bandcamp.com/album/through-sharp-eyes-we-witness" target="_blank">bandcamp</a><br>
+        <a href="https://linktr.ee/stasys" target="_blank">stasys' links</a><br><br>
+        <iframe width="100%" height="450" scrolling="no" frameborder="no" allow="autoplay" src="https://w.soundcloud.com/player/?url=https%3A//api.soundcloud.com/playlists/1731632490&color=%23ff5500&auto_play=false&hide_related=true&show_comments=true&show_user=true&show_reposts=false&show_teaser=false"></iframe><div style="font-size: 10px; color: #cccccc;line-break: anywhere;word-break: normal;overflow: hidden;white-space: nowrap;text-overflow: ellipsis; font-family: Interstate,Lucida Grande,Lucida Sans Unicode,Lucida Sans,Garuda,Verdana,Tahoma,sans-serif;font-weight: 100;"><a href="https://soundcloud.com/stasysmusial" title="Stasys" target="_blank" style="color: #cccccc; text-decoration: none;">Stasys</a> · <a href="https://soundcloud.com/stasysmusial/sets/through-sharp-eyes-we-witness" title="THROUGH SHARP EYES WE WITNESS by the modulated realm" target="_blank" style="color: #cccccc; text-decoration: none;">THROUGH SHARP EYES WE WITNESS by the modulated realm</a></div>
+
+
+        <br>
+        <p><b>credits:</b></p>
+        <p>
+            Stasys, Kyodaina, fwocl, myself, coil circuit, MANSTRETTA, Rhino Betatron, Viplex, Keitha, Fehu.
+        </p>
+    </body>
+</html>
