+++ content/coalition/data.json
@@ -1,0 +1,22 @@
+{
+    "title": "Coalition",
+    "dates": {
+        "yyyy-mm-dd": "published"
+    },
+    "categories": [
+        "music"
+    ],
+    "tags": [
+        "collab",
+        "megacollab",
+        "single"
+    ],
+    "links": {
+        "soundcloud": "",
+        "bandcamp": "",
+        "webpage": ""
+    },
+    "contributors": {
+        "Popbot": "primary creator"
+    }
+}
