+++ content/invention-for-tub/data.json
@@ -1,0 +1,20 @@
+{
+    "title": "Invention for Tub",
+    "dates": {
+        "yyyy-mm-dd": "published"
+    },
+    "categories": [
+        "music"
+    ],
+    "tags": [
+        "collab"
+    ],
+    "links": {
+        "soundcloud": "",
+        "bandcamp": "",
+        "webpage": ""
+    },
+    "contributors": {
+        "Popbot": "primary creator"
+    }
+}
