+++ content/ray-volpe---lazerbeam-(ragtime-edit)/data.json
@@ -1,0 +1,22 @@
+{
+    "title": "Ray Volpe - Lazerbeam (ragtime edit)",
+    "dates": {
+        "yyyy-mm-dd": "published"
+    },
+    "categories": [
+        "music"
+    ],
+    "tags": [
+        "single",
+        "remix",
+        "shitpost"
+    ],
+    "links": {
+        "soundcloud": "",
+        "webpage": ""
+    },
+    "contributors": {
+        "Popbot": "primary creator",
+        "Ray Volpe": "dubstep man"
+    }
+}
