+++ content/reactant/description.txt
@@ -1,1 +1,3 @@
-<h1>do this</h1>
+a 6-track EP I made over the summer of 2021. Unlike some of my other EPs, this has some amt of a theme, being the sort of bubble excited idm bass stuff.
+Making this taught me a lot about workflow and sound design, but I still had a lot to learn in terms of mixing and composition. A lot of it employed some seriously ignorant tactics. But that's ok, it's still fun for me to listen to from time to time.
+I especially enjoy listening to cascade. Also, i think that restraint is one of the first times i began experimenting with new textures that i wasn't used to hearing.
