+++ content/style.css
@@ -39,3 +39,9 @@
     padding: 0px;
     vertical-align: top;
 }
+
+.error {
+    color: rgb(255, 196, 196);
+    font-family: 'Courier New', Courier, monospace;
+    background-color: black;
+}
