+++ old/index.html
@@ -1,0 +1,103 @@
+<!DOCTYPE html>
+<!--[if lt IE 7]>      <html class="no-js lt-ie9 lt-ie8 lt-ie7"> <![endif]-->
+<!--[if IE 7]>         <html class="no-js lt-ie9 lt-ie8"> <![endif]-->
+<!--[if IE 8]>         <html class="no-js lt-ie9"> <![endif]-->
+<!--[if gt IE 8]>      <html class="no-js"> <!--<![endif]-->
+<html>
+    <head>
+        <meta charset="utf-8">
+        <meta http-equiv="X-UA-Compatible" content="IE=edge">
+        <title>popbot.work</title>
+        <meta name="description" content="">
+        <meta name="viewport" content="width=device-width, initial-scale=1">
+        <link rel="stylesheet" href="style.css">
+
+        <!-- DEFINING SVG FIILTERS FOR CSS -->
+        <svg height="0" width="0">
+            <defs>
+                <filter id="offset" width="180" height="180">
+                    <feOffset in="SourceGraphic" dx="0" dy="50">
+                        <animate 
+                            attributeName="dy"
+                            values="-50; 0; 50; 0; -50"
+                            dur="10s"
+                            repeatCount="indefinite"
+                        />
+                    </feOffset>              
+                </filter>
+                <filter id="hueRotateText">
+                    <feColorMatrix type="hueRotate" values="180" />
+                    <!-- <feOffset in="SourceGraphic" dx="00" dy="0" /> -->
+                </filter>
+            </defs>
+        </svg>
+        
+
+    </head>
+    <body>
+        <p id="mousecoords">😺😺😺🦊🦊🦊🦊rmmmrrp!</p>
+        <p>wip draft for a complete redesign of my  <a href="https://www.popbot.work/index.html">current website</a></p>
+        
+        <!-- MUSIC -->
+        <div id="container-music" data-category="music">
+            <center><b>music</b></center>
+            <div class="item">
+                <span class="year">year</span> | <span class="title">title</span> <span class="tags">tag(s)</span>
+            </div>
+            <hr>
+        </div>
+
+        <!-- CONTRAPTIONS -->
+        <div id="container-contraption" data-category="contrahttps://popbot11.github.io/website-v2/">
+            <center><b>contraptions</b></center>
+            <div class="item">
+                <span class="year">year</span> | <span class="title">title</span> <span class="tags">tag(s)</span>
+            </div>
+            <hr>
+        </div>
+
+        <!-- TAGLIST -->
+        <div id="container-taglist">
+            <center><span>Exclude:<br></span></center>
+            <hr>
+        </div>
+        
+        <div id="container-video">
+            <center><span>Video<br></span></center>
+        </div>
+
+        <div id="container-externallinks">
+            <center><span>My external links<br></span></center>
+        </div>
+
+        <div id="container-other">
+            <center><span>whatever other bullshit?<br></span></center>
+        </div>
+
+        <!-- METADATA -->
+        <div id="metadata">
+            <b>hello! welcomem to my page :)</b><br>
+            My name is Robin, and I make music and music-related things. Most people online know me as Popbot. 
+            This website is a place where I put all of the stuff I'm personally involved with, and probably other things too.<br>
+            I'm currently a student at Oberlin College & Conservatory, persuing a double degree in Computer Science and TIMARA (Technology in Music and Related Arts)</div>
+        <div id="description">description of entry will be here</div>
+
+        <!-- BULLSHIT -->
+        <button onclick="dripMode()" class="dripMode">enter Drip Mode</button>
+        <button onclick="turboMode()" class="turboMode">enter Turbo Mode</button>
+        <a  href="expertmode.html"><button class="crazyMode">enter Expert Mode</button></a>
+        
+        <div class="the88x31s">
+            <a href="https://minor-locrian.neocities.org/" target="_blank">
+                <img src="media/buttons/88x31javascriptfree.gif" >
+            </a>
+            <img src="media/buttons/88x31lies(false).gif"><br>
+            <img src="media/buttons/ben_barker_spectral.png">
+            <img src="media/buttons/popbotbutton.gif">
+        </div>
+    
+        <!-- <iframe width="100%" height="166" scrolling="no" frameborder="no" src="https://w.soundcloud.com/player/?url=https%3A//api.soundcloud.com/tracks/101276036&amp;color=ff6600&amp;auto_play=false&amp;show_artwork=true"></iframe> -->
+    
+        <script src="script.js" type="module"></script>        
+    </body>
+</html>
